github_pr: a tool to handle github pull requests
A tool to handle github pull requests.
It is a replacement of the github-status in the repo:
  https://github.com/SUSE-Cloud/automation/tree/master/scripts/github-status

The issues with github-status:

It is bound to the SUSE-Cloud organization and our few repos.
It could be changed to set status for other repos, but all defaults
were SUSE and cloud specific.

Due to the need to call github-status several times within a jenkins
job, we wasted several API requests and hit the rate limit several
times.

Other teams had interest in using github-status, but had to modify
it to meet their needs.

github_pr fixes these issues:
- it is configurable (org, repo, filters, actions)
- it will login once and cache data to save API requests
- it can call external actions during the filtering
- it brings a set of useful filters and actions
- it can be extended with team/site specific filters and actions

The idea of github_pr is:
For a given github organization and repositories it reads a filter
definition from a config file. This definition describes a filter
chain. PRs that do not match the filter are discarded, the others
are handed to the next filter in the chain.
After each filter actions can be run on the PRs that match and
that do not match (before they are discarded).

This will allow to replace the several pr trigger jobs with a single
jenkins job that does the polling which saves even more API requests.