github_pr: fix hash keys for create_status function
The keys of the parameter hash for create_status need
to be strings. This is due to the fact that the parameters
might also come from a json config wich will result in
string values as keys.

When this was adapted it the parameter hash creation
for the 'set-status' action was not changed accordingly.

This PR should fix the issue by changing the keys to strings
instead of symbols.