insert filter to chain for just one PR
The new parameter --this supports to insert filter
at the beginning of the filter chain. This is useful
when triggering actions or dryruns manually.
The use case may be to get the rebuild URLs or to execute
the blacklist or whitelist actions from the config for
just one pull request.

The parameter --this takes a github URL with the format:
  https://github.com/<org>/<repo>/pull/<pr_id>/commits/<sha1>

So in order to trigger the URL of the latest commit of this
pull request has to be passed. Otherwise the filter will
not match. This is small protection to make sure, the person
who triggered the run really saw the changes.