support for inverse filters
The new setting "inverse: true" in a filter config section
allows to invert the filter. This can help with filters that
are not easily negatable in the yaml config, like the ruby
regular expressions of the file matching filter.

Put differently:
With this it is now possible to write a 'positive' filter rule
then set it to "inverse: true" and add actions for the matches
to the blacklist_handler, while the other pulls are still
passed on to the next filter in the filter chain.
This is needed for the file matching filter.